AI-and-Analytics/Features-and-Functionality/IntelTensorFlow_AMX_BF16_Training/requirements.txt
absl-py
astunparse
h5py
huggingface-hub
keras
Keras-Preprocessing
matplotlib
numpy
pandas
Pillow
py-cpuinfo
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tf-estimator-nightly
tokenizers
transformers
